refactor: Rename BufTextWriter to Console, moving to top-level
