+++ kernel/src/graphics/text.rs
@@ -1,7 +1,5 @@
-use core::fmt::Write;
-
 use crate::graphics::fonts::{Font, Shinonome};
-use crate::graphics::{Canvas, Color, Colors, Position};
+use crate::graphics::{Canvas, Color, Position};
 
 const FONT_HEIGHT: usize = 16;
 const FONT_WIDTH: usize = 8;
@@ -48,45 +46,3 @@ pub(crate) trait TextWriter<'a>: Canvas<'a> {
 }
 
 impl<'a, T> TextWriter<'a> for T where T: Canvas<'a> {}
-
-pub(crate) struct BufTextWriter<'a, W> {
-    writer: &'a mut W,
-    position: Position,
-    color: Color,
-}
-
-impl<'a, W> BufTextWriter<'a, W> {
-    pub(crate) fn new(writer: &'a mut W) -> Self {
-        Self {
-            writer,
-            position: Position::zero(),
-            color: Colors::black(),
-        }
-    }
-
-    pub(crate) fn with_position<P>(mut self, position: P) -> Self
-    where
-        P: Into<Position>,
-    {
-        self.position = position.into();
-        self
-    }
-
-    pub(crate) fn with_color<C>(mut self, color: C) -> Self
-    where
-        C: Into<Color>,
-    {
-        self.color = color.into();
-        self
-    }
-}
-
-impl<'a, W> Write for BufTextWriter<'a, W>
-where
-    W: TextWriter<'a>,
-{
-    fn write_str(&mut self, s: &str) -> core::fmt::Result {
-        self.position = self.writer.write_string(self.position, s, self.color);
-        Ok(())
-    }
-}
