+++ kernel/src/main.rs
@@ -4,15 +4,18 @@
 #![feature(type_alias_impl_trait)]
 #![feature(generic_associated_types)]
 
+mod console;
 mod graphics;
 
-use core::fmt::Write;
 #[cfg(not(test))]
 use core::panic::PanicInfo;
+
+use core::fmt::Write;
 use mikan_core::KernelArgs;
 
+use crate::console::Console;
 use crate::graphics::frame_buffer::FrameBuffer;
-use crate::graphics::text::{BufTextWriter, TextWriter};
+use crate::graphics::text::TextWriter;
 use crate::graphics::{Canvas, Colors, Region};
 
 #[panic_handler]
@@ -31,11 +34,11 @@ extern "C" fn kernel_main(args: KernelArgs) -> ! {
     frame_buffer.write_chars((0, 50).into(), '!'..='~', Colors::black());
     frame_buffer.write_string((0, 66).into(), "Hello, world!", Colors::blue());
 
-    let mut w = BufTextWriter::new(&mut frame_buffer)
+    let mut c = Console::new(&mut frame_buffer)
         .with_position((0, 82))
         .with_color(Colors::black());
 
-    writeln!(w, "1 + 2 = {}", 1 + 2).ok();
+    writeln!(c, "1 + 2 = {}", 1 + 2).ok();
 
     loop {
         aarch64::instructions::halt();
